v8r1 - fix for gcc3.2
git-svn-id: http://svn.cern.ch/guest/lhcb/Gauss/trunk@90883 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 005fc03e52872ec64b3154f34cba51e6f9298a99

+++ Gauss/Sim/GaussTools/src/Components/TrCutsRunAction.h
@@ -66,7 +66,7 @@ private:
   double m_ncut;
   double m_mcut;
   bool m_killloops;
-  double m_maxsteps;
+  int m_maxsteps;
   double m_minstep;
 };
 // ============================================================================
